TF05/api/requirements.txt
flask
flask-cors
pyyaml
requests
mysql-connector-python